+++ fugitive-darkness-models/src/main/java/com/zer0s2m/fugitivedarkness/models/GitRepoModel.java
@@ -27,21 +27,26 @@ public class GitRepoModel {
     @JsonProperty("is_load")
     private boolean isLoad;
 
+    @JsonProperty("is_local")
+    private boolean isLocal;
+
     public GitRepoModel() {}
 
-    public GitRepoModel(String group, String project, String host, Path source, boolean isLoad) {
+    public GitRepoModel(String group, String project, String host, Path source, boolean isLoad, boolean isLocal) {
         this.group = group;
         this.project = project;
         this.host = host;
         this.source = source.toString();
         this.isLoad = isLoad;
+        this.isLocal = isLocal;
     }
 
-    public GitRepoModel(String group, String project, String host, Path source) {
+    public GitRepoModel(String group, String project, String host, Path source, boolean isLocal) {
         this.group = group;
         this.project = project;
         this.host = host;
         this.isLoad = false;
+        this.isLocal = isLocal;
         this.source = source.toString();
     }
 
@@ -73,11 +78,15 @@ public class GitRepoModel {
         return source;
     }
 
+    public boolean getIsLocal() {
+        return isLocal;
+    }
+
     public String toString() {
         return String.format(
-                "%s[id=%s,group=%s,project=%s,host=%s,createdAt=%s,isLoad=%s,source=%s]",
+                "%s[id=%s,group=%s,project=%s,host=%s,createdAt=%s,isLoad=%s,source=%s, isLocal=%s]",
                 super.toString(),
-                id, group, project, host, createdAt, isLoad, source
+                id, group, project, host, createdAt, isLoad, source, isLocal
         );
     }