+++ fugitive-darkness-models/src/main/resources/db/migration/V1_5_0__add_column_is_local_in_git_repo.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "git_repositories"
+    ADD COLUMN is_local BOOLEAN NOT NULL DEFAULT FALSE;